Move test_fixauth_dryrun_has_feedback
module it was in was marked as smoke
This test uses temp_appliance fixture, and cannot be run in smoke testing
